Backend/langgraph_agents/requirements.txt
langchain-openai>=0.0.5
langchain>=0.1.0
langgraph>=0.0.10
openai>=1.3.7
python-dotenv>=1.0.0
aiohttp>=3.8.0
azure-identity>=1.12.0
azure-core>=1.26.4 
langchain-openai 
langchain-community
pymongo